cmd/note_common: move noteRunFn to the common code
noteRunFn() is being used as entry code for both mr_note and issue_note,
therefore it makes more sense to move this function to note_common.go
file instead of let it in issue_note.go.
